backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Werkzeug
PyJWT

